Add documentation and type hints
